// src/services/rest-api/customer-service/model/get-fatca-w9-output.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Customer Service API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.0.267
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface GetFatcaW9Output
 */
export interface GetFatcaW9Output {
    /**
     * 
     * @type {string}
     * @memberof GetFatcaW9Output
     */
    'name'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof GetFatcaW9Output
     */
    'businessName'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof GetFatcaW9Output
     */
    'appropriate'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof GetFatcaW9Output
     */
    'address1'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof GetFatcaW9Output
     */
    'address2'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof GetFatcaW9Output
     */
    'address3'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof GetFatcaW9Output
     */
    'zipCode'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof GetFatcaW9Output
     */
    'accountList'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof GetFatcaW9Output
     */
    'ssn'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof GetFatcaW9Output
     */
    'ein'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof GetFatcaW9Output
     */
    'exemptionPayeeCode'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof GetFatcaW9Output
     */
    'exemptionReportingCode'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof GetFatcaW9Output
     */
    'refId'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof GetFatcaW9Output
     */
    'refType'?: string | null;
}